--- infra/app/apim-mcp/mcp-prm.policy.xml
@@ -1,55 +1,0 @@
-<!--
-    Policy for Protected Resource Metadata endpoint (RFC 9728)
-    This endpoint provides authorization information and should be accessible without authentication
--->
-<policies>
-    <inbound>
-        <base />
-        <!-- No authentication required for PRM endpoint -->
-        
-        <!-- Return PRM response directly without calling backend -->
-        <return-response>
-            <set-status code="200" reason="OK" />
-            <set-header name="Content-Type" exists-action="override">
-                <value>application/json</value>
-            </set-header>
-            <set-header name="Cache-Control" exists-action="override">
-                <value>public, max-age=3600</value>
-            </set-header>
-            <set-body>@{
-                var gatewayUrl = "{{APIMGatewayURL}}";
-                var tenantId = "{{McpTenantId}}";
-                var clientId = "{{McpClientId}}";
-                
-                var prm = new JObject
-                {
-                    ["resource"] = gatewayUrl,
-                    ["authorization_servers"] = new JArray
-                    {
-                        new JObject
-                        {
-                            ["authorization_endpoint"] = $"https://login.microsoftonline.com/{tenantId}/oauth2/v2.0/authorize",
-                            ["token_endpoint"] = $"https://login.microsoftonline.com/{tenantId}/oauth2/v2.0/token",
-                            ["scopes_supported"] = new JArray { $"api://{clientId}/user_impersonate" },
-                            ["response_types_supported"] = new JArray { "code" },
-                            ["grant_types_supported"] = new JArray { "authorization_code" },
-                            ["token_endpoint_auth_methods_supported"] = new JArray { "none" },
-                            ["code_challenge_methods_supported"] = new JArray { "S256" }
-                        }
-                    }
-                };
-                
-                return prm.ToString();
-            }</set-body>
-        </return-response>
-    </inbound>
-    <backend>
-        <base />
-    </backend>
-    <outbound>
-        <base />
-    </outbound>
-    <on-error>
-        <base />
-    </on-error>
-</policies>
