--- infra/core/apim/apim.bicep
@@ -1,90 +1,0 @@
-/**
- * @module apim
- * @description This module defines the Azure API Management (APIM) resources using Bicep.
- * It includes configurations for creating and managing APIM instance.
- */
-
-// ------------------
-//    PARAMETERS
-// ------------------
-
-@description('The name of the API Management instance.')
-param apiManagementName string
-
-@description('The location of the API Management instance. Defaults to the resource group location.')
-param location string = resourceGroup().location
-
-@description('The email address of the publisher. Defaults to "noreply@microsoft.com".')
-param publisherEmail string = 'noreply@microsoft.com'
-
-@description('The name of the publisher. Defaults to "Microsoft".')
-param publisherName string = 'Microsoft'
-
-@description('Name of the APIM Logger')
-param apimLoggerName string = 'apim-logger'
-
-@description('Description of the APIM Logger')
-param apimLoggerDescription string = 'APIM Logger for Weather MCP Server'
-
-@description('The pricing tier of this API Management service')
-@allowed([
-  'Consumption'
-  'Developer'
-  'Basic'
-  'Basicv2'
-  'Standard'
-  'Standardv2'
-  'Premium'
-])
-param apimSku string = 'Basicv2'
-
-@description('The instrumentation key for Application Insights')
-param appInsightsInstrumentationKey string = ''
-
-@description('The resource ID for Application Insights')
-param appInsightsId string = ''
-
-@description('Tags to apply to the resources')
-param tags object = {}
-
-// ------------------
-//    RESOURCES
-// ------------------
-
-// https://learn.microsoft.com/azure/templates/microsoft.apimanagement/service
-resource apimService 'Microsoft.ApiManagement/service@2024-06-01-preview' = {
-  name: apiManagementName
-  location: location
-  tags: tags
-  sku: {
-    name: apimSku
-    capacity: 1
-  }
-  properties: {
-    publisherEmail: publisherEmail
-    publisherName: publisherName
-  }
-}
-
-// Create a logger only if we have an App Insights ID and instrumentation key.
-resource apimLogger 'Microsoft.ApiManagement/service/loggers@2021-12-01-preview' = if (!empty(appInsightsId) && !empty(appInsightsInstrumentationKey)) {
-  name: apimLoggerName
-  parent: apimService
-  properties: {
-    credentials: {
-      instrumentationKey: appInsightsInstrumentationKey
-    }
-    description: apimLoggerDescription
-    isBuffered: false
-    loggerType: 'applicationInsights'
-    resourceId: appInsightsId
-  }
-}
-
-// ------------------
-//    OUTPUTS
-// ------------------
-
-output id string = apimService.id
-output name string = apimService.name
-output gatewayUrl string = apimService.properties.gatewayUrl
