--- infra/core/identity/userAssignedIdentity.bicep
@@ -1,14 +1,0 @@
-param identityName string
-param location string
-param tags object = {}
-
-resource userAssignedIdentity 'Microsoft.ManagedIdentity/userAssignedIdentities@2023-07-31-preview' = {
-  name: identityName
-  location: location
-  tags: tags
-}
-
-output identityId string = userAssignedIdentity.id
-output identityName string = userAssignedIdentity.name
-output identityPrincipalId string = userAssignedIdentity.properties.principalId
-output identityClientId string = userAssignedIdentity.properties.clientId
