+++ mcp-handler/function.json
